--- mllib/src/main/scala/org/apache/spark/ml/linalg/distributed/DistributedMatrix.scala
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.ml.linalg.distributed
-
-import breeze.linalg.{DenseMatrix => BDM}
-
-import org.apache.spark.annotation.Since
-
-/**
- * Represents a distributively stored matrix backed by one or more RDDs.
- */
-@Since("1.0.0")
-trait DistributedMatrix extends Serializable {
-
-  /** Gets or computes the number of rows. */
-  @Since("1.0.0")
-  def numRows(): Long
-
-  /** Gets or computes the number of columns. */
-  @Since("1.0.0")
-  def numCols(): Long
-
-  /** Collects data and assembles a local dense breeze matrix (for test only). */
-  private[ml] def toBreeze(): BDM[Double]
-}
