+++ mllib/src/test/scala/org/apache/spark/mllib/linalg/distributed/BlockMatrixSuite.scala
@@ -304,10 +304,25 @@ class BlockMatrixSuite extends SparkFunSuite with MLlibTestSparkContext {
     val largeB = new BlockMatrix(sc.parallelize(largerBblocks, 8), 4, 4)
     val largeC = largeA.multiply(largeB)
     val localC = largeC.toLocalMatrix()
+    val largeC2 = largeA.multiply(largeB, 2, 2, 2, 2, 2)
+    val localC2 = largeC2.toLocalMatrix()
+    val largeC3 = largeA.multiply(largeB, 3, 2, 3, 3, 2)
+    val localC3 = largeC3.toLocalMatrix()
+    val largeC4 = largeA.multiply(largeB, 5, 4, 4, 3, 2)
+    val localC4 = largeC4.toLocalMatrix()
     val result = largeA.toLocalMatrix().multiply(largeB.toLocalMatrix().asInstanceOf[DenseMatrix])
     assert(largeC.numRows() === largeA.numRows())
     assert(largeC.numCols() === largeB.numCols())
+    assert(largeC2.numRows() === largeA.numRows())
+    assert(largeC2.numCols() === largeB.numCols())
+    assert(largeC3.numRows() === largeA.numRows())
+    assert(largeC3.numCols() === largeB.numCols())
+    assert(largeC4.numRows() === largeA.numRows())
+    assert(largeC4.numCols() === largeB.numCols())
     assert(localC ~== result absTol 1e-8)
+    assert(localC2 ~== result absTol 1e-8)
+    assert(localC3 ~== result absTol 1e-8)
+    assert(localC4 ~== result absTol 1e-8)
   }
 
   test("simulate multiply") {
@@ -318,7 +333,7 @@ class BlockMatrixSuite extends SparkFunSuite with MLlibTestSparkContext {
     val B = new BlockMatrix(rdd, colPerPart, rowPerPart)
     val resultPartitioner = GridPartitioner(gridBasedMat.numRowBlocks, B.numColBlocks,
       math.max(numPartitions, 2))
-    val (destinationsA, destinationsB) = gridBasedMat.simulateMultiply(B, resultPartitioner)
+    val (destinationsA, destinationsB) = gridBasedMat.simulateMultiply(B, resultPartitioner, 1)
     assert(destinationsA((0, 0)) === Set(0))
     assert(destinationsA((0, 1)) === Set(2))
     assert(destinationsA((1, 0)) === Set(0))
